+++ core/docs/changelog.md
@@ -1,4 +1,7 @@
 # TBD
+### Fixes
+* Improved error reporting for typos in service ids to pause/unpause service
+
 ### Changes
 * Switch the API container to storing its data in the enclave data volume, rather than in the bindmounted enclave data dirpath on the host machine